Make I/O queue requests more compact, use u64 instead of u64 for some structures
This should have better compatibility with ABIs such as x32

+++ kernel/src/scheduler/process/mod.rs
@@ -24,7 +24,7 @@ pub struct Process {
 }
 
 struct PendingTicket {
-	user_data: usize,
+	user_data: u64,
 	data_ptr: *mut u8,
 	data_len: usize,
 	ticket: TicketOrJob,
